+++ novawallet/Modules/WalletsList/Common/WalletIconView.swift
@@ -1,7 +1,7 @@
 import UIKit
 import SnapKit
 
-final class ProxyIconView: UIView {
+final class WalletIconView: UIView {
     enum Constants {
         static let holeWidth: CGFloat = 3
         static let iconSize = CGSize(width: 32, height: 32)
